requirements.txt
MouseInfo==0.1.3
msgpack==1.0.0
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.2.2
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycrypto==2.6.1
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.3
